# Avoid heap allocations in Htonl