--- a/pkg/inet.go
+++ b/pkg/inet.go
@@ -1,35 +1,17 @@
 package rpmdb
 
 import (
-	"bytes"
 	"encoding/binary"
-	"log"
 )
 
 func Htonl(val int32) int32 {
-	buf := new(bytes.Buffer)
-	if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
-		log.Println(err)
-		return 0
-	}
-
-	if err := binary.Read(buf, binary.BigEndian, &val); err != nil {
-		log.Println(err)
-		return 0
-	}
-	return val
+	var buf [4]byte
+	binary.LittleEndian.PutUint32(buf[:], uint32(val))
+	return int32(binary.BigEndian.Uint32(buf[:]))
 }
 
 func HtonlU(val uint32) uint32 {
-	buf := new(bytes.Buffer)
-	if err := binary.Write(buf, binary.LittleEndian, val); err != nil {
-		log.Println(err)
-		return 0
-	}
-
-	if err := binary.Read(buf, binary.BigEndian, &val); err != nil {
-		log.Println(err)
-		return 0
-	}
-	return val
+	var buf [4]byte
+	binary.LittleEndian.PutUint32(buf[:], val)
+	return binary.BigEndian.Uint32(buf[:])
 }