Avoids heap allocations when in Htonl.

```
rpmdb_test.go:131: Allocations per run "CentOS5 plain": from 296182 to 26664
rpmdb_test.go:131: Allocations per run "CentOS6 Plain": from 346814 to 41826
rpmdb_test.go:131: Allocations per run "CentOS6 with Development tools": from 703604 to 82604
rpmdb_test.go:131: Allocations per run "CentOS6 with many packages": from 866345 to 96344
rpmdb_test.go:131: Allocations per run "CentOS7 Plain": from 389679 to 46470
rpmdb_test.go:131: Allocations per run "CentOS7 with Development tools": from 721811 to 100046
rpmdb_test.go:131: Allocations per run "CentOS7 with many packages": from 1061689 to 127268
rpmdb_test.go:131: Allocations per run "CentOS7 with Python 3.5": from 911443 to 106971
rpmdb_test.go:131: Allocations per run "CentOS7 with httpd 2.4": from 588830 to 62563
rpmdb_test.go:131: Allocations per run "CentOS8 with modules": from 1440866 to 136506
rpmdb_test.go:131: Allocations per run "RHEL UBI8 from s390x": from 508158 to 51291
rpmdb_test.go:131: Allocations per run "SLE15 with NDB style rpm database": from 98312 to 5654
rpmdb_test.go:131: Allocations per run "Fedora35 with SQLite3 style rpm database": from 371913 to 22497
rpmdb_test.go:131: Allocations per run "Fedora35 plus MongoDB with SQLite3 style rpm database": from 382463 to 23151
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.