@jinroh 2024-11-25
Before

```
goos: darwin
goarch: arm64
pkg: github.com/knqyf263/go-rpmdb/pkg
cpu: Apple M1 Max
BenchmarkRpmDB_Package/CentOS5_plain-10                                       52          23991929 ns/op        49760552 B/op     296190 allocs/op
BenchmarkRpmDB_Package/CentOS6_Plain-10                                       40          28327141 ns/op        62770936 B/op     346915 allocs/op
BenchmarkRpmDB_Package/CentOS6_with_Development_tools-10                      18          60692363 ns/op        169853252 B/op    703798 allocs/op
BenchmarkRpmDB_Package/CentOS6_with_many_packages-10                          16          69095055 ns/op        190209496 B/op    866544 allocs/op
BenchmarkRpmDB_Package/CentOS7_Plain-10                                       37          31174939 ns/op        74387664 B/op     389787 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_Development_tools-10                      15          74731719 ns/op        240852056 B/op    721985 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_many_packages-10                          12          90756823 ns/op        276267332 B/op   1061907 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_Python_3.5-10                             14          79331298 ns/op        227163886 B/op    911661 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_httpd_2.4-10                              28          41661811 ns/op        97779868 B/op     588991 allocs/op
BenchmarkRpmDB_Package/CentOS8_with_modules-10                                12          91337743 ns/op        234611008 B/op   1441177 allocs/op
BenchmarkRpmDB_Package/RHEL_UBI8_from_s390x-10                                32          35893936 ns/op        94996768 B/op     508319 allocs/op
BenchmarkRpmDB_Package/SLE15_with_NDB_style_rpm_database-10                  255           4641358 ns/op         6810925 B/op      98331 allocs/op
BenchmarkRpmDB_Package/Fedora35_with_SQLite3_style_rpm_database-10                    55          20780058 ns/op        33609021 B/op     371972 allocs/op
BenchmarkRpmDB_Package/Fedora35_plus_MongoDB_with_SQLite3_style_rpm_database-10                       55          25275911 ns/op        34199078 B/op     382543 allocs/op
PASS
```

------------

AFTER

```
goos: darwin
goarch: arm64
pkg: github.com/knqyf263/go-rpmdb/pkg
cpu: Apple M1 Max
BenchmarkRpmDB_Package/CentOS5_plain-10                                       66          18810401 ns/op        41674131 B/op      26672 allocs/op
BenchmarkRpmDB_Package/CentOS6_Plain-10                                       50          22820472 ns/op        53618285 B/op      41945 allocs/op
BenchmarkRpmDB_Package/CentOS6_with_Development_tools-10                      22          49735311 ns/op        151215652 B/op     82787 allocs/op
BenchmarkRpmDB_Package/CentOS6_with_many_packages-10                          20          55669265 ns/op        167104224 B/op     96553 allocs/op
BenchmarkRpmDB_Package/CentOS7_Plain-10                                       46          25464075 ns/op        64089220 B/op      46576 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_Development_tools-10                      18          64353773 ns/op        222201363 B/op    100239 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_many_packages-10                          15          74720281 ns/op        248222750 B/op    127488 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_Python_3.5-10                             16          65440023 ns/op        203025041 B/op    107189 allocs/op
BenchmarkRpmDB_Package/CentOS7_with_httpd_2.4-10                              34          32086971 ns/op        81990270 B/op      62729 allocs/op
BenchmarkRpmDB_Package/CentOS8_with_modules-10                                18          67958331 ns/op        195470964 B/op    136830 allocs/op
BenchmarkRpmDB_Package/RHEL_UBI8_from_s390x-10                                40          28148409 ns/op        81290482 B/op      51460 allocs/op
BenchmarkRpmDB_Package/SLE15_with_NDB_style_rpm_database-10                  453           2649647 ns/op         4029938 B/op       5671 allocs/op
BenchmarkRpmDB_Package/Fedora35_with_SQLite3_style_rpm_database-10                    76          13913482 ns/op        23125151 B/op      22573 allocs/op
BenchmarkRpmDB_Package/Fedora35_plus_MongoDB_with_SQLite3_style_rpm_database-10                       90          13983450 ns/op        23415763 B/op      23217 allocs/op
PASS
```